+++ programs/depin_home/src/state/config.rs
@@ -5,7 +5,7 @@ use anchor_lang::prelude::*;
 pub struct Config {
     pub admin: Pubkey,
     pub rewards_bump: u8,
-    pub reward_amount_temp: u16,
-    pub reward_amount_noise: u16,
+    pub reward_amount_temp: u8,
+    pub reward_amount_noise: u8,
     pub bump: u8,
 }
