+++ programs/depin_home/src/state/user_config.rs
@@ -4,7 +4,7 @@ use anchor_lang::prelude::*;
 #[derive(InitSpace)]
 pub struct UserConfig {
     pub points: u32,
-    pub temp_data_points: u64,
-    pub noise_data_points: u64,
+    pub temp_data_points: u32,
+    pub noise_data_points: u32,
     pub bump: u8,
 }
